refactor: remove profile service
