+++ src/app/features/auth/components/change-password-dialog/change-password-dialog/change-password-dialog.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { POLYMORPHEUS_CONTEXT } from '@taiga-ui/polymorpheus';
 import { of } from 'rxjs';
 import { provideHttpClient } from '@angular/common/http';
-import { ProfileService } from '@features/auth/services/profile/profile.service';
+import { ProfileApiService } from '@features/auth/services/profile-api/profile-api.service';
 import { ChangePasswordDialogComponent } from './change-password-dialog.component';
 
 describe('ChangePasswordDialogComponent', () => {
@@ -22,7 +22,7 @@ describe('ChangePasswordDialogComponent', () => {
           provide: POLYMORPHEUS_CONTEXT,
           useValue: { data: of({}) },
         },
-        { provide: ProfileService, useValue: profileServiceMock },
+        { provide: ProfileApiService, useValue: profileServiceMock },
       ],
     }).compileComponents();
