+++ src/app/features/auth/pages/profile-page/profile-page.component.ts
@@ -3,10 +3,14 @@ import { TuiButton, TuiAutoColorPipe, TuiDialogService } from '@taiga-ui/core';
 import { TuiAvatar } from '@taiga-ui/kit';
 import { PolymorpheusComponent } from '@taiga-ui/polymorpheus';
 import { AuthService } from '@features/auth/services/auth/auth.service';
-import { ProfileService } from '@features/auth/services/profile/profile.service';
 import { AsyncPipe, NgIf } from '@angular/common';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { FormControl } from '@angular/forms';
+import { AlertService } from '@core/services/alert/alert.service';
+import { Router } from '@angular/router';
+import { ProfileApiService } from '@features/auth/services/profile-api/profile-api.service';
+import { LocalStorageService } from '@core/services/local-storage/local-storage.service';
+import { LocalStorageKey } from '@shared/enums/local-storage-key.enum';
+import { tap } from 'rxjs';
 import { ProfileFieldComponent } from '../../components/profile-field/profile-field.component';
 import { ChangePasswordDialogComponent } from '../../components/change-password-dialog/change-password-dialog/change-password-dialog.component';
 
@@ -27,9 +31,11 @@ import { ChangePasswordDialogComponent } from '../../components/change-password-
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProfilePageComponent {
-  private readonly authService = inject(AuthService);
+  private readonly profileApiService = inject(ProfileApiService);
 
-  private readonly profileService = inject(ProfileService);
+  private readonly localStorage = inject(LocalStorageService);
+
+  private readonly authService = inject(AuthService);
 
   private readonly dialogs = inject(TuiDialogService);
 
@@ -37,23 +43,43 @@ export class ProfilePageComponent {
 
   private readonly destroy = inject(DestroyRef);
 
-  public control = new FormControl('');
+  private readonly alert = inject(AlertService);
 
-  private readonly dialog = this.dialogs.open<number>(
-    new PolymorpheusComponent(ChangePasswordDialogComponent, this.injector),
-    {
+  private readonly router = inject(Router);
+
+  private readonly dialog = this.dialogs
+    .open<number>(new PolymorpheusComponent(ChangePasswordDialogComponent, this.injector), {
       dismissible: true,
       label: 'Change Password',
-    }
+    })
+    .pipe(takeUntilDestroyed());
+
+  protected userInformation = this.profileApiService.getUserInformation().pipe(
+    takeUntilDestroyed(this.destroy),
+    tap({
+      error: ({ error: { message } }) => {
+        this.alert.open({ message, label: 'Error', appearance: 'error' });
+        this.router.navigate(['/home']);
+      },
+    })
   );
 
-  protected userInformation = this.profileService.getUserInformation();
-
   protected logout() {
-    this.authService.logout().pipe(takeUntilDestroyed(this.destroy)).subscribe();
+    this.authService
+      .logout()
+      .pipe(takeUntilDestroyed(this.destroy))
+      .subscribe({
+        next: () => {
+          this.router.navigate(['/home']);
+          this.localStorage.removeItem(LocalStorageKey.UserToken);
+        },
+        error: ({ error: { message } }) => {
+          this.alert.open({ message, label: 'Error', appearance: 'error' });
+        },
+      });
   }
 
   protected showDialog() {
-    this.dialog.pipe(takeUntilDestroyed(this.destroy)).subscribe();
+    this.dialog.subscribe();
   }
 }
