--- src/app/features/auth/services/profile/profile.service.spec.ts
@@ -1,17 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { provideHttpClient } from '@angular/common/http';
-import { ProfileService } from './profile.service';
-
-describe('ProfileService', () => {
-  let service: ProfileService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({ providers: [provideHttpClient()] });
-    service = TestBed.inject(ProfileService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
